Relocate usage of context types
