+++ packages/moleculer-graphql/src/classes/SchemaBuilder.ts
@@ -4,8 +4,8 @@ import { stitchingDirectives } from '@graphql-tools/stitching-directives';
 import type { IFieldResolver, IResolvers } from '@graphql-tools/utils';
 import type { GraphQLSchema } from 'graphql';
 import type { Service } from 'moleculer';
+import type { GraphQLContext } from '../functions';
 import { buildFullActionName, ensureArray } from '../utils';
-import type { GraphQLContext } from '.';
 
 export type SchemaDirectiveTransformer = (schema: GraphQLSchema) => GraphQLSchema;
