+++ packages/moleculer-graphql/src/classes/index.ts
@@ -1,8 +1,4 @@
 export { default as GatewayStitcher } from './GatewayStitcher';
-export {
-	default as GraphQLExecutor,
-	type GraphQLContextFactory,
-	type GraphQLContext,
-} from './GraphQLExecutor';
+export { default as GraphQLExecutor } from './GraphQLExecutor';
 export { default as RequestHandler, type Request } from './RequestHandler';
 export { default as SchemaBuilder, type SchemaDirectiveTransformer } from './SchemaBuilder';
