+++ packages/moleculer-graphql/src/functions/index.ts
@@ -1,2 +1,7 @@
-export { default as createGraphQLContext } from './createGraphQLContext';
+export {
+	default as createGraphQLContext,
+	type GraphQLContextFactory,
+	type GraphQLContextBase,
+	type GraphQLContext,
+} from './createGraphQLContext';
 export { default as createValidate } from './createValidate';
