+++ packages/moleculer-graphql/src/mixins/gatewayMixin.ts
@@ -5,7 +5,8 @@ import { defaultsDeep } from 'lodash';
 import type { Service, ServiceSchema } from 'moleculer';
 import type { Route } from 'moleculer-web';
 import { GatewayStitcher, RequestHandler } from '../classes';
-import type { GraphQLContextFactory, Request } from '../classes';
+import type { Request } from '../classes';
+import type { GraphQLContextFactory } from '../functions';
 
 interface GatewayService<TGraphQLContext extends Record<string, unknown>> extends Service {
 	rebuildSchema: boolean;
